Changed Camera weight to float

+++ Assets/Scripts/Network/NetworkCameraController.cs
@@ -75,7 +75,7 @@ public class NetworkCameraController : NetworkBehaviour
         private void _setCameraFOV()
         {
             Vector3 total = Vector3.zero;
-            int length = 0;
+            float length = 0;
 
             foreach (CameraTargets ct in Context._cameraTargets)
             {
